// src/lib/validation/form.schema.ts
import { z } from 'zod';

export const loginSchema = z.object({
    email: z.string().email("Invalid email address"),
    password: z.string().min(6, "Password must be at least 6 characters long"),
})

export const registerSchema = z.object({
    name: z.string().min(1, 'Name is required'),
    email: z.string().email("Email address is required"),
    password: z.string().min(6, "Password must be at least 6 characters long"),
})

export const formSchemas = {
    login: loginSchema,
    register: registerSchema
}